Adição da funcionalidade Remover Veiculo e correção de bugs
