+++ ocupa.py
@@ -4,28 +4,6 @@ from datetime import datetime
 import pandas as pd
 from function import *
 
-def registerSaida(matricula):
-    index = 0
-    vehiclePark = False
-    with open('ocupacao.csv') as csv_file:
-        pandaReader = pd.read_csv(csv_file, sep = ';')
-        for row in pandaReader:
-            if (pandaReader.at[index, "3"] == "None" and pandaReader.at[index, "0"] == correctMatricula(matricula)):
-                pandaReader.iat[index, 3] = str(datetime.now())
-                pandaReader.iat[index, 4] = str(int(calcDurationMinutes(pandaReader.at[index, "2"], datetime.now())))
-                pandaReader.iat[index, 5] = str(int(calcDurationMinutes(pandaReader.at[index, "2"], datetime.now())) * 0.1)
-                vehiclePark = True
-                break
-            index += 1
-        pandaReader.to_csv("ocupacao.csv", index=False, sep = ';')
-    if (vehiclePark == False):
-        return "Veiculo não encontrado."
-
-def calcDurationMinutes(entrada, saida):
-    datetimeFormat = '%Y-%m-%d %H:%M:%S.%f'
-    diff = datetime.strptime(str(saida), datetimeFormat) - datetime.strptime(str(entrada), datetimeFormat)
-    return diff.seconds / 60
-
 class Ocupa:
 
     def __init__(self, veiculo, entrada = datetime.now(), saida= None):
@@ -76,7 +54,7 @@ class Ocupa:
         return self.__duracao
 
     def setVeiculo(self,l_veiculo):
-        if  isinstance(l_veiculo, Veiculo) == False or self.vehicleValidation(l_veiculo) == False:
+        if  isinstance(l_veiculo, Veiculo) == False or vehicleValidation(l_veiculo) == False:
             pass
         else:
             self.__veiculo = l_veiculo
@@ -93,14 +71,6 @@ class Ocupa:
         else:
             self.__saida=l_saida
 
-    def vehicleValidation(self, l_veiculo):
-        with open('ep1.csv') as csv_file:
-            reader = csv.reader(csv_file, delimiter = ';')
-            for row in reader:
-                if str(row[0]) == str(l_veiculo.getMatricula()):
-                    return True
-            return False
-
     def __repr__(self):
         return {'Veiculo':self.getVeiculo(), 'Entrada': self.getEntrada().strftime("%m/%d/%Y, %H:%M:%S")}
 
@@ -112,3 +82,5 @@ class Ocupa:
             if(self.getEntrada() == other.getEntrada() and self.getVeiculo() == other.getVeiculo()):
                 return True
         return False
+
+registerSaida("00aa00")
